* 更新测试结算

+++ afc-api/src/main/java/rebue/afc/vpay/ro/VpayNotifyRo.java
@@ -1,5 +1,6 @@
 package rebue.afc.vpay.ro;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
@@ -13,35 +14,35 @@ public class VpayNotifyRo {
     /**
      * 用户ID
      */
-    private Long   userId;
+    private Long       userId;
     /**
      * 支付账户ID
      */
-    private String payAccountId;
+    private String     payAccountId;
     /**
      * 支付交易的金额(单位为元)
      */
-    private Double payAmount;
+    private BigDecimal payAmount;
     /**
      * 支付返现金的金额(单位为元)
      */
-    private Double payChangeAmount1;
+    private BigDecimal payChangeAmount1;
     /**
      * 支付余额的金额(单位为元)
      */
-    private Double payChangeAmount2;
+    private BigDecimal payChangeAmount2;
     /**
      * 支付订单号
      */
-    private String payOrderId;
+    private String     payOrderId;
     /**
      * 订单号
      */
-    private String orderId;
+    private String     orderId;
     /**
      * 支付完成时间
      */
-    private Date   payTime;
+    private Date       payTime;
 
     public Long getUserId() {
         return userId;
@@ -59,27 +60,27 @@ public class VpayNotifyRo {
         this.payAccountId = payAccountId;
     }
 
-    public Double getPayAmount() {
+    public BigDecimal getPayAmount() {
         return payAmount;
     }
 
-    public void setPayAmount(Double payAmount) {
+    public void setPayAmount(BigDecimal payAmount) {
         this.payAmount = payAmount;
     }
 
-    public Double getPayChangeAmount1() {
+    public BigDecimal getPayChangeAmount1() {
         return payChangeAmount1;
     }
 
-    public void setPayChangeAmount1(Double payChangeAmount1) {
+    public void setPayChangeAmount1(BigDecimal payChangeAmount1) {
         this.payChangeAmount1 = payChangeAmount1;
     }
 
-    public Double getPayChangeAmount2() {
+    public BigDecimal getPayChangeAmount2() {
         return payChangeAmount2;
     }
 
-    public void setPayChangeAmount2(Double payChangeAmount2) {
+    public void setPayChangeAmount2(BigDecimal payChangeAmount2) {
         this.payChangeAmount2 = payChangeAmount2;
     }
